refactor(watch): extract useWatchSuppressionSync (sync + hold-expire effects)
