+++ src/renderer/shared/hooks/watch/index.ts
@@ -12,4 +12,5 @@ export * from "./useWatchPing";
 export * from "./watchEngine";
 export * from "./useWatchEngine";
 export * from "./useWatchSessionMeta";
+export * from "./useWatchSuppressionSync";
 export * from "./watchStallRecovery";
